move tests to subfolder
