+++ tests/Mscc.GenerativeAI/AIPlatform_Samples_Should.cs
