+++ tests/Mscc.GenerativeAI/GoogleAi_GeminiProVision_Should.cs
