+++ tests/Mscc.GenerativeAI/GoogleAi_GeminiPro_Should.cs
