+++ tests/Mscc.GenerativeAI/Test.Mscc.GenerativeAI.csproj
@@ -63,8 +63,8 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\src\Mscc.GenerativeAI\Mscc.GenerativeAI.csproj" />
-		<ProjectReference Include="..\src\Mscc.GenerativeAI.Web\Mscc.GenerativeAI.Web.csproj" />
+		<ProjectReference Include="..\..\src\Mscc.GenerativeAI\Mscc.GenerativeAI.csproj" />
+<!--		<ProjectReference Include="..\..\src\Mscc.GenerativeAI.Web\Mscc.GenerativeAI.Web.csproj" />-->
 	</ItemGroup>
 
 	<ProjectExtensions><VisualStudio><UserProperties appsettings_1json__JsonSchema="" /></VisualStudio></ProjectExtensions>
