+++ tests/Mscc.GenerativeAI/TestExtensions.cs
