+++ tests/Mscc.GenerativeAI/VertexAi_Gemini15Pro_Should.cs
