+++ tests/Mscc.GenerativeAI/VertexAi_GeminiProVision_Should.cs
