+++ tests/Mscc.GenerativeAI/VertexAi_GeminiPro_Should.cs
