+++ tests/Mscc.GenerativeAI/appsettings.json
