+++ tests/Mscc.GenerativeAI/payload/cat.jpg
