+++ tests/Mscc.GenerativeAI/payload/scones.jpg
