+++ tests/Mscc.GenerativeAI/xunit.runner.json
